# docs/openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/logtest":{"get":{"summary":"Read Root","operationId":"read_root_logtest_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/token":{"post":{"summary":"Login","operationId":"login_token_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_token_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Healthcheck","operationId":"healthcheck__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/users/me":{"get":{"tags":["users"],"summary":"Read Users Me","operationId":"read_users_me_users_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/users/":{"get":{"tags":["users"],"summary":"Read Users","operationId":"read_users_users__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/User"},"type":"array","title":"Response Read Users Users  Get"}}}}}},"post":{"tags":["users"],"summary":"Create User","operationId":"create_user_users__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/deactivate":{"post":{"tags":["users"],"summary":"Deactivate User","operationId":"deactivate_user_users_deactivate_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/users/activate":{"post":{"tags":["users"],"summary":"Activate User","operationId":"activate_user_users_activate_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"username","in":"query","required":true,"schema":{"type":"string","title":"Username"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/storage/":{"get":{"tags":["storage"],"summary":"Read Files","operationId":"read_files_storage__get","security":[{"OAuth2PasswordBearer":[]}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FileOverview"}}}}}},"delete":{"tags":["storage"],"summary":"Delete Object","operationId":"delete_object_storage__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"path","in":"query","required":true,"schema":{"type":"string","title":"Path"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"anyOf":[{"$ref":"#/components/schemas/FileSummary"},{"$ref":"#/components/schemas/DirectorySummary"}]},"title":"Response Delete Object Storage  Delete"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/storage/file":{"post":{"tags":["storage"],"summary":"Upload Object","operationId":"upload_object_storage_file_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"path","in":"query","required":true,"schema":{"type":"string","title":"Path"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_object_storage_file_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FileSummary"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/storage/directory":{"post":{"tags":["storage"],"summary":"Create Directory","operationId":"create_directory_storage_directory_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DirectoryCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DirectorySummary"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/storage/download":{"post":{"tags":["storage"],"summary":"Download Object","operationId":"download_object_storage_download_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"path","in":"query","required":true,"schema":{"type":"string","title":"Path"}}],"responses":{"200":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/groups/groups":{"get":{"tags":["groups"],"summary":"Read Groups","operationId":"read_groups_groups_groups_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Group"},"type":"array","title":"Response Read Groups Groups Groups Get"}}}}}}},"/groups/me":{"get":{"tags":["groups"],"summary":"Read My Groups","operationId":"read_my_groups_groups_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Group"},"type":"array","title":"Response Read My Groups Groups Me Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/groups/":{"post":{"tags":["groups"],"summary":"Create Group","operationId":"create_group_groups__post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"group_name","in":"query","required":true,"schema":{"type":"string","title":"Group Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Group"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/groups/members":{"post":{"tags":["groups"],"summary":"Create Group","operationId":"create_group_groups_members_post","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"group_name","in":"query","required":true,"schema":{"type":"string","title":"Group Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/User"},"title":"Response Create Group Groups Members Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/groups/join":{"patch":{"tags":["groups"],"summary":"Add User To Group","operationId":"add_user_to_group_groups_join_patch","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"group_name","in":"query","required":true,"schema":{"type":"string","title":"Group Name"}},{"name":"user_name","in":"query","required":true,"schema":{"type":"string","title":"User Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Group"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_login_token_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"password"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_token_post"},"Body_upload_object_storage_file_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_object_storage_file_post"},"Directory":{"properties":{"name":{"type":"string","title":"Name"},"path":{"type":"string","title":"Path"},"owner":{"anyOf":[{"$ref":"#/components/schemas/User"},{"type":"null"}]},"permission":{"anyOf":[{"$ref":"#/components/schemas/Permission"},{"type":"null"}]},"children":{"items":{"anyOf":[{"$ref":"#/components/schemas/File"},{"$ref":"#/components/schemas/Directory"}]},"type":"array","title":"Children"}},"type":"object","required":["name","path","children"],"title":"Directory"},"DirectoryCreate":{"properties":{"path":{"type":"string","title":"Path"},"name":{"type":"string","title":"Name"}},"type":"object","required":["path","name"],"title":"DirectoryCreate"},"DirectorySummary":{"properties":{"name":{"type":"string","title":"Name"},"path":{"type":"string","title":"Path"}},"type":"object","required":["name","path"],"title":"DirectorySummary"},"DirectorySummaryChildren":{"properties":{"name":{"type":"string","title":"Name"},"path":{"type":"string","title":"Path"},"children":{"items":{"anyOf":[{"$ref":"#/components/schemas/DirectorySummaryChildren"},{"$ref":"#/components/schemas/FileSummary"}]},"type":"array","title":"Children"}},"type":"object","required":["name","path","children"],"title":"DirectorySummaryChildren"},"File":{"properties":{"name":{"type":"string","title":"Name"},"path":{"type":"string","title":"Path"},"owner":{"anyOf":[{"$ref":"#/components/schemas/User"},{"type":"null"}]},"permission":{"anyOf":[{"$ref":"#/components/schemas/Permission"},{"type":"null"}]},"filetype":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Filetype"}},"type":"object","required":["name","path"],"title":"File"},"FileOverview":{"properties":{"owned_objects":{"$ref":"#/components/schemas/DirectorySummaryChildren"},"shared_objects":{"items":{"anyOf":[{"$ref":"#/components/schemas/Directory"},{"$ref":"#/components/schemas/File"}]},"type":"array","title":"Shared Objects"},"group_shared_objects":{"additionalProperties":{"items":{"anyOf":[{"$ref":"#/components/schemas/Directory"},{"$ref":"#/components/schemas/File"}]},"type":"array"},"type":"object","title":"Group Shared Objects"}},"type":"object","required":["owned_objects","shared_objects","group_shared_objects"],"title":"FileOverview"},"FileSummary":{"properties":{"name":{"type":"string","title":"Name"},"path":{"type":"string","title":"Path"},"filetype":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Filetype"}},"type":"object","required":["name","path"],"title":"FileSummary"},"Group":{"properties":{"name":{"type":"string","title":"Name"},"members":{"items":{"$ref":"#/components/schemas/User"},"type":"array","title":"Members"}},"type":"object","required":["name","members"],"title":"Group"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Permission":{"type":"string","enum":["R","RW"],"title":"Permission"},"Token":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type"},"quota":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Quota"},"used":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Used"}},"type":"object","required":["access_token","token_type"],"title":"Token"},"User":{"properties":{"username":{"type":"string","title":"Username"},"is_active":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Is Active"},"quota":{"type":"integer","title":"Quota"},"used":{"type":"integer","title":"Used"}},"type":"object","required":["username","quota","used"],"title":"User"},"UserCreate":{"properties":{"username":{"type":"string","title":"Username"},"is_active":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Is Active"},"password":{"type":"string","title":"Password"}},"type":"object","required":["username","password"],"title":"UserCreate"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"token"}}}}}}